by andrewbelcher: Moved user creation trait to make it clear it's for tests only.
